requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
plotly
python-dotenv
langchain
langchain_core
langchain_community
langchain_google_genai